platform-checks: Fix further permission problems with RBAC
