+++ misc/python/materialize/checks/owners.py
@@ -17,12 +17,17 @@ from materialize.util import MzVersion
 class Owners(Check):
     def _create_objects(self, role: str, i: int, expensive: bool = False) -> str:
         s = dedent(
-            f"""
+            (
+                f"""
             $ postgres-execute connection=postgres://mz_system@materialized:6877/materialize
             GRANT CREATE ON DATABASE materialize TO {role}
             GRANT CREATE ON SCHEMA materialize.public TO {role}
             GRANT CREATE ON CLUSTER default TO {role}
-
+            """
+                if self.current_version >= MzVersion.parse("0.52.0-dev")
+                else ""
+            )
+            + f"""
             $ postgres-execute connection=postgres://{role}@materialized:6875/materialize
             CREATE DATABASE owner_db{i}
             CREATE SCHEMA owner_schema{i}
