+++ test/upgrade-matrix/nodes.py
@@ -10,7 +10,7 @@
 from typing import List, Optional
 
 from materialize.checks.actions import Action, Initialize, Manipulate, Validate
-from materialize.checks.mzcompose_actions import KillMz, StartMz
+from materialize.checks.mzcompose_actions import ConfigureMz, KillMz, StartMz
 from materialize.checks.scenarios import Scenario
 from materialize.util import MzVersion
 
@@ -40,7 +40,10 @@ class BeginVersion(Node):
     def actions(self, scenario: Scenario) -> List[Action]:
         # As this action may need start very old Mz versions,
         # we do not use any bootstrap_systme_parameters
-        return [StartMz(tag=self.version, bootstrap_system_parameters=[])]
+        return [
+            StartMz(tag=self.version, bootstrap_system_parameters=[]),
+            ConfigureMz(scenario),
+        ]
 
 
 class EndVersion(Node):
